prov/sockets: Move socket provider code out of generic windows osd file
sock_get_ip_addr_table is a socket specific call.  It should be part
of the socket provider, not part of generic windows osd code.
